feat: :sparkles: added more information in the anime detailed page
