+++ src/renderer/graphql/queries/animeQueries/AnimeInfo.graphql
@@ -1,32 +1,7 @@
 query getAnimeInfo($mediaId: Int) {
   Media(id: $mediaId, type: ANIME) {
     id
-    countryOfOrigin
-    season
-    seasonYear
-    startDate {
-      year
-      month
-      day
-    }
-    endDate {
-      year
-      month
-      day
-    }
-    studios(isMain: true) {
-      nodes {
-        id
-        name
-        siteUrl
-      }
-    }
     description
-    trailer {
-      site
-      id
-    }
-    synonyms
     relations {
       edges {
         relationType(version: 2)
@@ -86,30 +61,10 @@ query getAnimeInfo($mediaId: Int) {
           coverImage {
             large
           }
-        }
-      }
-    }
-  }
-  Page(page: 1) {
-    pageInfo {
-      total
-      perPage
-      currentPage
-      lastPage
-      hasNextPage
-    }
-    mediaList(isFollowing: true, sort: [STATUS], mediaId: $mediaId) {
-      id
-      status
-      score(format: POINT_100)
-      progress
-      progressVolumes
-      user {
-        id
-        name
-        avatar {
-          large
-          medium
+          mediaListEntry {
+            progress
+            score(format: POINT_100)
+          }
         }
       }
     }
